[rust] move `(Model|Motion)IOPluginController` to its own namespace
