+++ rust/plugin_wasm/src/model/mod.rs
@@ -5,6 +5,7 @@
 */
 
 mod c_api;
+mod controller;
 mod core;
 mod plugin;
