+++ rust/plugin_wasm/src/motion/mod.rs
@@ -5,6 +5,7 @@
 */
 
 mod c_api;
+mod controller;
 mod core;
 mod plugin;
