+++ rust/plugin_wasm/src/motion/test/full.rs
@@ -11,7 +11,7 @@ use pretty_assertions::assert_eq;
 use serde_json::json;
 
 use crate::motion::{
-    plugin::MotionIOPluginController,
+    controller::MotionIOPluginController,
     test::{create_random_data, inner_create_controller, read_plugin_output, Pipe, PluginOutput},
 };
